customercarebe/requirements.txt
asgiref==3.8.1
av==15.0.0
beautifulsoup4==4.13.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
ctranslate2==4.6.0
cycler==0.12.1
deepface==0.0.93
Django==4.2.18
django-bootstrap-v5==1.0.11
django-cors-headers==4.9.0
django-phonenumber-field==8.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.54.1
fsspec==2025.3.2
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
keyring==25.6.0
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numpy==2.1.2
onnxruntime==1.22.1
packaging==24.1
pillow==11.0.0
protobuf==6.31.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.9.0
six==1.16.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.6
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
